# lib/api-zod/src/generated/types/activityLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityLogMetadata } from "./activityLogMetadata";

export interface ActivityLog {
  id: number;
  session_id: string;
  /** @nullable */
  ip_address?: string | null;
  /** @nullable */
  user_agent?: string | null;
  action_type: string;
  /** @nullable */
  metadata?: ActivityLogMetadata;
  flagged: boolean;
  /** @nullable */
  flag_reason?: string | null;
  /** @nullable */
  flagged_by?: string | null;
  /** @nullable */
  flagged_at?: string | null;
  created_at: string;
}