# lib/api-zod/src/generated/types/activityLogsPage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityLog } from "./activityLog";

export interface ActivityLogsPage {
  items: ActivityLog[];
  total: number;
  page: number;
  limit: number;
}