# lib/api-zod/src/generated/types/activityStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityStats {
  total_sessions: number;
  total_events: number;
  flagged_sessions: number;
  consents_today: number;
  events_today: number;
}