# lib/api-zod/src/generated/types/flagActivityBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FlagActivityBody {
  flag_reason: string;
  /** @nullable */
  flagged_by?: string | null;
}