# lib/api-zod/src/generated/types/logActivityBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LogActivityBodyMetadata } from "./logActivityBodyMetadata";

export interface LogActivityBody {
  session_id: string;
  action_type: string;
  /** @nullable */
  metadata?: LogActivityBodyMetadata;
}