# lib/api-zod/src/generated/types/recordConsentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RecordConsentBody {
  session_id: string;
  /** @nullable */
  user_agent?: string | null;
  consented_items: string[];
}