# lib/api-zod/src/generated/types/sessionStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SessionStatus {
  session_id: string;
  flagged: boolean;
  /** @nullable */
  flag_reason?: string | null;
  /** @nullable */
  flagged_at?: string | null;
}